Move esp.web.util.template to esp.utils.cache_inclusion_tag
esp.utils.template was already taken, and the file only contains one thing
anyway.
