+++ esp/esp/miniblog/templatetags/render_blog.py
@@ -1,7 +1,7 @@
 
 
 from django import template
-from esp.web.util.template import cache_inclusion_tag
+from esp.utils.cache_inclusion_tag import cache_inclusion_tag
 
 register = template.Library()
